backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos MongoDB
motor==3.3.2
pymongo==4.6.0

# Modelos y validación
pydantic==2.5.2
pydantic[email]==2.5.0
pydantic[email]==2.5.2

# Variables de entorno
python-dotenv==1.0.0

# Análisis de datos 
pandas==2.1.4

# SDK de MCP para Python
mcp[cli]
# Pydantic settings (requerido por MCP y config)
pydantic-settings>=2.0.0


anthropic>=0.34.0
openai